feat(pgsql-seed): add batteries-included seeding package with pgpm integration
- Create new pgsql-seed package that re-exports everything from pg-seed
- Add deployPgpm/loadPgpm functions for pgpm package deployment
- Update pgsql-test to depend on pgsql-seed instead of pg-seed
- Update pg-seed README to mention pgsql-seed for pgpm integration
- pgsql-seed is the public-facing package, pg-seed is the lightweight core
