+++ postgres/pgsql-seed/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "pgsql-test",
-  "version": "2.18.17",
+  "name": "pgsql-seed",
+  "version": "0.1.0",
   "author": "Constructive <developers@constructive.io>",
-  "description": "pgsql-test offers isolated, role-aware, and rollback-friendly PostgreSQL environments for integration tests — giving developers realistic test coverage without external state pollution",
+  "description": "PostgreSQL seeding utilities with pgpm integration - batteries included",
   "main": "index.js",
   "module": "esm/index.js",
   "types": "index.d.ts",
@@ -22,28 +22,15 @@
   "keywords": [
     "postgres",
     "postgresql",
-    "testing",
-    "integration-tests",
-    "database-testing",
-    "pg",
-    "rls",
-    "role-based-access",
-    "test-database",
-    "test-runner",
-    "jest",
-    "mocha",
-    "sqitch",
-    "constructive",
-    "graphile",
-    "typeorm",
-    "knex",
     "seed",
-    "fixtures",
-    "transactions",
-    "rollback",
-    "node-postgres",
-    "pg-pool",
-    "pg-client"
+    "seeding",
+    "csv",
+    "json",
+    "sql",
+    "pgpm",
+    "migrations",
+    "data-loading",
+    "constructive"
   ],
   "scripts": {
     "clean": "makage clean",
@@ -56,18 +43,13 @@
   },
   "devDependencies": {
     "@types/pg": "^8.16.0",
-    "@types/pg-copy-streams": "^1.2.5",
     "makage": "^0.1.9"
   },
   "dependencies": {
     "@pgpmjs/core": "workspace:^",
     "@pgpmjs/env": "workspace:^",
-    "@pgpmjs/logger": "workspace:^",
-    "@pgpmjs/server-utils": "workspace:^",
-    "@pgpmjs/types": "workspace:^",
     "pg": "^8.16.3",
-    "pg-cache": "workspace:^",
     "pg-env": "workspace:^",
-    "pgsql-seed": "workspace:^"
+    "pg-seed": "workspace:^"
   }
 }
