+++ postgres/pgsql-test/src/seed/json.ts
@@ -1,11 +1,11 @@
-// Re-export from pg-seed for backwards compatibility
+// Re-export from pgsql-seed for backwards compatibility
 export {
   insertJson,
   insertJsonMap,
   type JsonSeedMap
-} from 'pg-seed';
+} from 'pgsql-seed';
 
-import { insertJsonMap, type JsonSeedMap } from 'pg-seed';
+import { insertJsonMap, type JsonSeedMap } from 'pgsql-seed';
 
 import { SeedAdapter, SeedContext } from './types';
