+++ postgres/pgsql-test/src/seed/sql.ts
@@ -1,6 +1,6 @@
-// Re-export from pg-seed for backwards compatibility
+// Re-export from pgsql-seed for backwards compatibility
 export {
   execSql,
   loadSql,
   loadSqlFiles
-} from 'pg-seed';
+} from 'pgsql-seed';
